+++ examples/misc/add_truthset_data.py
@@ -7,13 +7,14 @@ from senzing_truthset import (
 
 from . import sz_abstract_factory
 
+record_sets = [
+    TRUTHSET_CUSTOMER_RECORDS,
+    TRUTHSET_REFERENCE_RECORDS,
+    TRUTHSET_WATCHLIST_RECORDS,
+]
+
 try:
     sz_engine = sz_abstract_factory.create_engine()
-    record_sets = [
-        TRUTHSET_CUSTOMER_RECORDS,
-        TRUTHSET_REFERENCE_RECORDS,
-        TRUTHSET_WATCHLIST_RECORDS,
-    ]
     for record_set in record_sets:
         for record in record_set.values():
             sz_engine.add_record(