+++ examples/szconfigmanager/set_default_config.py
@@ -4,19 +4,20 @@ from senzing import SzError
 
 from . import sz_configmanager
 
+config_comment = "Just an example"
+data_source_code = f"REPLACE_DEFAULT_CONFIG_ID_{time.time()}"
+
 try:
 
     # Create a new config.
 
     sz_config = sz_configmanager.create_config_from_template()
-    data_source_code = f"REPLACE_DEFAULT_CONFIG_ID_{time.time()}"
     sz_config.add_data_source(data_source_code)
 
     # Persist the new default config.
 
-    CONFIG_DEFINITION = sz_config.export()
-    CONFIG_COMMENT = "Just an example"
-    CONFIG_ID = sz_configmanager.set_default_config(CONFIG_DEFINITION, CONFIG_COMMENT)
-
+    config_definition = sz_config.export()
+    config_id = sz_configmanager.set_default_config(config_definition, config_comment)
+    print(config_id)
 except SzError as err:
     print(f"\nERROR: {err}\n")