+++ examples/szengine/count_redo_records.py
@@ -3,7 +3,7 @@ from senzing import SzError
 from . import sz_engine
 
 try:
-    RESULT = sz_engine.count_redo_records()
-    print(f"\n{RESULT}\n")
+    result = sz_engine.count_redo_records()
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")