+++ examples/szengine/export_json_fetch_close.py
@@ -2,14 +2,15 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+flags = SzEngineFlags.SZ_EXPORT_DEFAULT_FLAGS
+
 try:
-    flags = SzEngineFlags.SZ_EXPORT_DEFAULT_FLAGS
-    EXPORT_HANDLE = sz_engine.export_json_entity_report(flags)
+    export_handle = sz_engine.export_json_entity_report(flags)
     while True:
-        FRAGMENT = sz_engine.fetch_next(EXPORT_HANDLE)
-        if not FRAGMENT:
+        fragment = sz_engine.fetch_next(export_handle)
+        if not fragment:
             break
-        print(FRAGMENT, end="")
-    sz_engine.close_export(EXPORT_HANDLE)
+        print(fragment, end="")
+    sz_engine.close_export(export_handle)
 except SzError as err:
     print(f"\nERROR: {err}\n")