+++ examples/szengine/find_network_by_entity_id.py
@@ -2,13 +2,16 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+build_out_degrees = 1
+build_out_max_entities = 10
+entity_list = [1, 4]
+flags = SzEngineFlags.SZ_FIND_NETWORK_DEFAULT_FLAGS
+max_degrees = 2
+
 try:
-    BUILD_OUT_DEGREES = 1
-    entity_list = [1, 4]
-    flags = SzEngineFlags.SZ_FIND_NETWORK_DEFAULT_FLAGS
-    MAX_DEGREES = 2
-    MAX_ENTITIES = 10
-    RESULT = sz_engine.find_network_by_entity_id(entity_list, MAX_DEGREES, BUILD_OUT_DEGREES, MAX_ENTITIES, flags)
-    print(f"\n{RESULT}\n")
+    result = sz_engine.find_network_by_entity_id(
+        entity_list, max_degrees, build_out_degrees, build_out_max_entities, flags
+    )
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")