+++ examples/szengine/find_network_by_record_id.py
@@ -2,13 +2,17 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+build_out_degrees = 1
+build_out_max_entities = 10
+flags = SzEngineFlags.SZ_FIND_NETWORK_DEFAULT_FLAGS
+max_degrees = 2
+max_entities = 10
+record_list = [("CUSTOMERS", "1001"), ("CUSTOMERS", "1009")]
+
 try:
-    BUILD_OUT_DEGREES = 1
-    flags = SzEngineFlags.SZ_FIND_NETWORK_DEFAULT_FLAGS
-    MAX_DEGREES = 2
-    MAX_ENTITIES = 10
-    record_list = [("CUSTOMERS", "1001"), ("CUSTOMERS", "1009")]
-    RESULT = sz_engine.find_network_by_record_id(record_list, MAX_DEGREES, BUILD_OUT_DEGREES, MAX_ENTITIES, flags)
-    print(f"\n{RESULT}\n")
+    result = sz_engine.find_network_by_record_id(
+        record_list, max_degrees, build_out_degrees, build_out_max_entities, flags
+    )
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")