+++ examples/szengine/find_path_by_entity_id.py
@@ -4,21 +4,22 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+avoid_entity_ids: List[int] = []
+end_entity_id = 400215
+flags = SzEngineFlags.SZ_FIND_PATH_DEFAULT_FLAGS
+max_degrees = 2
+required_data_sources: List[str] = []
+start_entity_id = 1
+
 try:
-    avoid_entity_ids: List[int] = []
-    END_ENTITY_ID = 4
-    flags = SzEngineFlags.SZ_FIND_PATH_DEFAULT_FLAGS
-    MAX_DEGREES = 2
-    required_data_sources: List[str] = []
-    START_ENTITY_ID = 1
-    RESULT = sz_engine.find_path_by_entity_id(
-        START_ENTITY_ID,
-        END_ENTITY_ID,
-        MAX_DEGREES,
+    result = sz_engine.find_path_by_entity_id(
+        start_entity_id,
+        end_entity_id,
+        max_degrees,
         avoid_entity_ids,
         required_data_sources,
         flags,
     )
-    print(f"\n{RESULT}\n")
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")