+++ examples/szengine/search_by_attributes.py
@@ -4,11 +4,11 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+attributes = json.dumps({"NAME_FULL": "Bob Smith", "EMAIL_ADDRESS": "bsmith@work.com"})
+flags = SzEngineFlags.SZ_SEARCH_BY_ATTRIBUTES_DEFAULT_FLAGS
+
 try:
-    attributes = json.dumps({"NAME_FULL": "BOB SMITH", "EMAIL_ADDRESS": "bsmith@work.com"})
-    flags = SzEngineFlags.SZ_SEARCH_BY_ATTRIBUTES_DEFAULT_FLAGS
-    SEARCH_PROFILE = ""
-    RESULT = sz_engine.search_by_attributes(attributes, flags, SEARCH_PROFILE)
-    print(f"\n{RESULT}\n")
+    result = sz_engine.search_by_attributes(attributes, flags)
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")