+++ examples/szengine/why_entities.py
@@ -2,15 +2,16 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+entity_id_1 = 1
+entity_id_2 = 400215
+flags = SzEngineFlags.SZ_WHY_ENTITIES_DEFAULT_FLAGS
+
 try:
-    ENTITY_ID_1 = 1
-    ENTITY_ID_2 = 4
-    flags = SzEngineFlags.SZ_WHY_ENTITIES_DEFAULT_FLAGS
-    RESULT = sz_engine.why_entities(
-        ENTITY_ID_1,
-        ENTITY_ID_2,
+    result = sz_engine.why_entities(
+        entity_id_1,
+        entity_id_2,
         flags,
     )
-    print(f"\n{RESULT}\n")
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")