+++ examples/szengine/why_record_in_entity.py
@@ -2,15 +2,16 @@ from senzing import SzEngineFlags, SzError
 
 from . import sz_engine
 
+data_source_code = "CUSTOMERS"
+flags = SzEngineFlags.SZ_WHY_RECORD_IN_ENTITY_DEFAULT_FLAGS
+record_id = "1001"
+
 try:
-    DATA_SOURCE_CODE = "CUSTOMERS"
-    flags = SzEngineFlags.SZ_WHY_RECORDS_DEFAULT_FLAGS
-    RECORD_ID = "1001"
-    RESULT = sz_engine.why_record_in_entity(
-        DATA_SOURCE_CODE,
-        RECORD_ID,
+    result = sz_engine.why_record_in_entity(
+        data_source_code,
+        record_id,
         flags,
     )
-    print(f"\n{RESULT}\n")
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")