+++ examples/szproduct/get_license.py
@@ -3,7 +3,7 @@ from senzing import SzError
 from . import sz_product
 
 try:
-    RESULT = sz_product.get_license()
-    print(f"\n{RESULT}\n")
+    result = sz_product.get_license()
+    print(f"\n{result}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")